--- .envrc
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-# this makes direnv go. direnv is good.
-
-# because ... they need it for some reason?
-export CI=true
-
-# you need TAURI_PRIVATE_KEY and APPLE_SIGNING_IDENTITY in .tauri.direnv to make Apple signing work
-source_up_if_exists .tauri.direnv
