--- .github/workflows/tauri_build.yml
@@ -1,39 +1,0 @@
----
-name: Build Tauri App
-"on":
-  pull_request:
-  push:
-
-jobs:
-  build:
-    strategy:
-      fail-fast: false
-      matrix:
-        platform: [macos-latest, ubuntu-20.04, windows-latest]
-
-    runs-on: ${{ matrix.platform }}
-    steps:
-      - uses: actions/checkout@v3
-      # - name: setup node
-      #   uses: actions/setup-node@v3
-      #   with:
-      #     node-version: 16
-      - name: Install Rust
-        uses: actions-rs/toolchain@v1
-        with:
-          toolchain: stable
-          override: true
-          default: true
-      - name: install dependencies (ubuntu only)
-        if: matrix.platform == 'ubuntu-20.04'
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y libgtk-3-dev libwebkit2gtk-4.0-dev \
-            libappindicator3-dev librsvg2-dev patchelf
-      # - name: install app dependencies and build it
-      #   run: yarn && yarn build
-      - name: Build dependencies
-        run: cargo build --workspace
-      - uses: tauri-apps/tauri-action@v0
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
