+++ .gitignore
@@ -1,4 +1,3 @@
-/dist/
 /target/
 /Cargo.lock
 .DS_Store
