--- .taurignore
@@ -1,3 +1,0 @@
-/src
-/public
-/Cargo.toml
