--- .vscode/extensions.json
@@ -1,3 +1,0 @@
-{
-  "recommendations": ["tauri-apps.tauri-vscode", "rust-lang.rust-analyzer"]
-}
