--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-  "emmet.includeLanguages": {
-    "rust": "html"
-  }
-}
