+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "memetool-ui"
+name = "memetool"
 version = "0.0.1"
 edition = "2021"
 
@@ -7,49 +7,12 @@ edition = "2021"
 [dependencies]
 eframe = "0.21.3"
 egui_extras = { version = "0.21.0", features = ["svg", "image"] }
-gloo = "0.8.0"
-image = {workspace = true}
 itertools = "0.10.5"
-js-sys = { workspace = true }
 lazy_static = "1.4.0"
-memetool_shared = { workspace = true }
-mime_guess = { workspace = true }
-poll-promise = "0.2.0"
 rayon = "1.7.0"
-serde = { workspace = true }
-serde-wasm-bindgen = { workspace = true }
-shellexpand = { workspace = true }
 tokio = { version = "1.27.0", features = ["sync", "full"] }
-wasm-bindgen = { workspace = true }
-wasm-bindgen-futures = { workspace = true }
-web-sys = { workspace = true }
-yew = { version="0.20.0", features = ["csr"] }
-yew-hooks = "0.2.0"
-yew-router = "0.17.0"
-
-[workspace.dependencies]
 base64 = "0.21.0"
 image = "0.24.5"
-js-sys = "0.3.59"
-memetool_shared = { path = "./memetool_shared" }
-mime_guess = "2.0.4"
 serde = { version = "1.0.140", features = ["derive"] }
 serde_json = { version = "1.0" }
-serde-wasm-bindgen = "0.4.3"
 shellexpand = "3.0.0"
-tauri = { version = "1.2", features = [
-    "config-json5",
-    "fs-read-dir",
-    "fs-read-file",
-    "icon-png",
-    "shell-open"
-    ]}
-wasm-bindgen = { version = "0.2.82", features = ["serde-serialize"] }
-wasm-bindgen-futures = "0.4.32"
-web-sys = "0.3.59"
-
-[workspace]
-members = [
-    "src-tauri",
-    "memetool_shared",
-]
