--- index.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title></title>
-    <link data-trunk rel="css" href="style.css" />
-    <link data-trunk rel="copy-dir" href="public" />
-  </head>
-</html>
