--- memetool_shared/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "memetool_shared"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-base64 = { workspace = true }
-image = { workspace = true }
-mime_guess = { workspace = true }
-serde = { workspace = true }
-
-
