--- src-tauri/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "memetool"
-version = "0.0.2"
-description = "MemeTool"
-authors = ["James Hodgkinson <james@terminaloutcomes.com>"]
-license = "MIT"
-repository = "https://github.com/yaleman/memetool/"
-edition = "2021"
-rust-version = "1.65"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[build-dependencies]
-tauri-build = { version = "1.2", features = [] }
-
-[dependencies]
-image = { workspace = true}
-memetool_shared = { workspace = true }
-serde = { workspace=true, features = ["derive"] }
-serde_json = { workspace=true }
-shellexpand = { workspace = true }
-tauri = { workspace = true , features = ["dialog-confirm", "fs-read-dir", "fs-read-file", "fs-remove-file", "protocol-asset", "shell-open"] }
-tokio = { version = "1.23.0", features = ["full"] }
-
-[features]
-# by default Tauri runs in production mode
-# when `tauri dev` runs it is executed with `cargo run --no-default-features` if `devPath` is an URL
-default = ["custom-protocol"]
-# this feature is used used for production builds where `devPath` points to the filesystem
-# DO NOT remove this
-custom-protocol = ["tauri/custom-protocol"]
