--- src-tauri/tauri.conf.json
@@ -1,88 +1,0 @@
-{
-  "build": {
-    "beforeDevCommand": "trunk build",
-    "devPath": "../dist",
-    "distDir": "../dist",
-    "withGlobalTauri": true
-  },
-  "package": {
-    "productName": "memetool",
-    "version": "0.0.0"
-  },
-  "tauri": {
-    "allowlist": {
-      "all": false,
-      "dialog" : {
-        "confirm" : true
-      },
-      "fs" : {
-        "readFile" : true,
-        "readDir" : true,
-        "removeFile": true,
-        "scope" : [
-          "*"
-        ]
-      },
-      "globalShortcut": {
-        "all": false
-      },
-      "protocol": {
-        "all": false,
-        "asset": true,
-        "assetScope": [
-          "**"
-        ]
-      },
-      "shell": {
-        "all": false,
-        "open": true
-      }
-    },
-
-    "bundle": {
-      "active": true,
-      "category": "Productivity",
-      "copyright": "James Hodgkinson 2022-Now",
-      "deb": {
-        "depends": []
-      },
-      "externalBin": [],
-      "icon": [
-        "icons/app-icon.icns",
-        "icons/app-icon.ico",
-        "icons/apple-touch-icon-base.png"
-      ],
-      "identifier": "com.terminaloutcomes.memetool",
-      "longDescription": "Memetool allows you to mess with your memes.",
-      "macOS": {
-        "entitlements": null,
-        "exceptionDomain": "",
-        "frameworks": [],
-        "providerShortName": null,
-        "signingIdentity": null,
-        "minimumSystemVersion": "12.0"
-      },
-      "resources": [
-      ],
-      "shortDescription": "Meme image management.",
-      "targets": "all",
-      "windows": {
-        "certificateThumbprint": null,
-        "digestAlgorithm": "sha256",
-        "timestampUrl": ""
-      }
-    },
-    "security": {
-      "csp": null
-    },
-    "windows": [
-      {
-        "fullscreen": false,
-        "height": 600,
-        "resizable": true,
-        "title": "MemeTool",
-        "width": 800
-      }
-    ]
-  }
-}
